+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v0.14.0](https://github.com/tshlabs/cadence/tree/0.14.0) - 2018-??-??
+* **Breaking change** - Rename the `MetricBuilder::send()` method to
+  `MetricBuilder::try_send()` and create a new `.send()` method that discards
+  successful results and invokes a custom handler for error results. Handlers
+  can be set by using the `StatsdClient::from_sink_with_handler` constructor.
+  Per [#65](https://github.com/tshlabs/cadence/issues/65).
+
 ## [v0.13.2](https://github.com/tshlabs/cadence/tree/0.13.2) - 2018-03-13
 * Warn when `MetricBuilder` instances aren't used when adding tags to metrics
   per [#63](https://github.com/tshlabs/cadence/issues/63).